+++ test_runner/src/main/kotlin/ftl/environment/NetworkConfigurationCatalog.kt
@@ -4,9 +4,9 @@ import ftl.environment.common.asPrintableTable
 import ftl.gc.GcTesting
 import ftl.http.executeWithRetry
 
-fun networkConfigurationAsTable() = networkConfiguration().asPrintableTable()
+fun networkConfigurationAsTable() = getNetworkConfiguration().asPrintableTable()
 
-private fun networkConfiguration() = GcTesting.get.testEnvironmentCatalog()
+fun getNetworkConfiguration() = GcTesting.get.testEnvironmentCatalog()
     .get("NETWORK_CONFIGURATION")
     .executeWithRetry()
     ?.networkConfigurationCatalog